refactor: move getLayout into file
